work on comments

+++ app/Http/Controllers/Admin/CommentController.php
@@ -34,45 +34,38 @@ class CommentController extends Controller
 
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
 
 
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
-        $content = Content::first();
 
+        $parent_id = (isset($request->parent_id)) ? $request->parent_id : null;
+
+        $model = app()->make($request->commentable_type);
+        $object = $model->findOrFail($request->commentable_id);
         $comment = new Comment();
-        $comment->message = str_random(500);
+        $comment->message = $request->message;
+        $comment->parent_id = $parent_id;
         $comment->user_id = auth()->user()->getKey();
-        $content->comments()->save($comment);
+        $object->comments()->save($comment);
+
+        session()->flash('success', __('messages.created_success'));
+
+        return back();
+
 
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Comment $comment
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function show(Comment $comment)
     {
-        //
+
     }
 
 
@@ -92,14 +85,7 @@ class CommentController extends Controller
         return view('admin.comments.edit', compact('comment'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param  \App\Comment             $comment
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, Comment $comment)
     {
         $comment->update($request->except(['_token', '_method']));
@@ -108,16 +94,14 @@ class CommentController extends Controller
         return redirect(route('admin.comments.index'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Comment $comment
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy(Comment $comment)
     {
-        $comment->delete();
-        return back()->with('success', __('messages.deleted_success'));
+        if($comment->hasChildren()){
+            return back()->with('error', __('messages.not_allowe_delete_child'));
+        }else{
+            $comment->delete();
+            return back()->with('success', __('messages.deleted_success'));
+        }
     }
 }
